Changed the database entries to have autoincrementing integer primary keys. Updated calls to the primary key to handle integers instead of longs. Updated main activity template to be more responsive.

+++ app/src/main/java/com/example/etaspare/stitchcounter/WriteToDb.java
@@ -32,7 +32,7 @@ public class WriteToDb extends AsyncTask<Counter, Void, Void> {
                 String strFilter = StitchCounterContract.CounterEntry._ID + "=" + values.get(StitchCounterContract.CounterEntry._ID);
                 db.update(StitchCounterContract.CounterEntry.TABLE_NAME, values, strFilter, null);
             } else {
-                counter[0].ID = db.insert(StitchCounterContract.CounterEntry.TABLE_NAME, null, values);
+                counter[0].ID = (int) db.insert(StitchCounterContract.CounterEntry.TABLE_NAME, null, values);
             }
         } finally {
             db.close();
